requirements.txt
h5py>=3.0.0
pillow>=9.0.0
numpy>=1.20.0
